requirements.txt
GitPython==2.1.7
gcloud==0.18.3
google-api-python-client==1.6.4
pyhocon==0.3.38
giturlparse.py==0.0.5
requests==2.18.4
pylint==1.7.4
oauth2client==4.1.2
clint==0.5.1
pex==1.2.13
wheel==0.29.0
